+++ src/core/camera.hpp
