+++ src/core/explosion.cpp
