+++ src/core/explosion.hpp
