+++ src/core/pixel.cpp
