+++ src/core/pixel.hpp
