+++ src/core/renderer.cpp
