+++ src/core/shape_renderer.cpp
