+++ src/core/texture.cpp
