+++ src/core/texture.hpp
