+++ src/core/update_rigid_bodies.cpp
