+++ src/core/utility.hpp
