+++ src/core/window.cpp
