+++ src/core/world.cpp
